Stability fixes for python_plugin_test
The "normal" timeout is eliminated. The "short" timeout is changed to
be the length used in tests that will time out as part of their
execution and the "long" timeout is changed to be a ridiculously high
value that will have no bearing on passing tests.

The "pause" behavior of _ServicerMethods is changed to use a
threading.Condition's wait/notify methods rather than busy-sleeping.

Tests that used servicer delay to verify that asynchronous calls are
not affected by server delay are changed to use servicer pause to
verify that asynchronous calls return while the servicer is paused.

Busy-sleeping in testHalfDuplexCallWedged is replaced with use of a
threading.Condition's wait/notify methods.

Fixes https://github.com/grpc/grpc/issues/1900.


Former-commit-id: 5ee11a8e1263eb00f1ddda0cd00570d9dd1681a0